Implement Request Tracking
When a request is made to a shortened URL, (i.e. `uri.to/abcd123`),
create a `requests` row and associate it with the MappedUrl.

This commit also introduces the `Statistics` model, which isn't backed
by the database. It is responsible for calculating statistics for each
URL. Currently, its only stat is a count of views, but it could track
popular referrers, most recent views, views by month, etc.